+++ nlsam/__init__.py
@@ -1,5 +1,6 @@
 import sys
 import os
+# import ctypes
 
 
 def fix_multiproc_windows():
@@ -37,6 +38,26 @@ def fix_multiproc_windows():
 
         # Second override 'Popen' class with our modified version.
         forking.Popen = _Popen
+
+
+def get_setup_params():
+    params = {}
+    params['scripts'] = ['scripts/nlsam_denoising']
+    params['name'] = 'nlsam'
+    params['author'] = 'Samuel St-Jean'
+    params['author_email'] = 'samuel@isi.uu.nl'
+    params['url'] = 'https://github.com/samuelstjean/nlsam'
+    params['version'] = '0.5.1'
+    params['install_requires'] = ['numpy>=1.10.4',
+                                  'scipy>=0.14',
+                                  'cython>=0.21',
+                                  'cythongsl>=0.2.1',
+                                  'nibabel>=2.0',
+                                  'spams>=2.4']
+    params['dependency_links'] = ['https://github.com/samuelstjean/spams-python/releases/download/0.1/spams-2.6.zip#egg=spams-2.6']
+
+    return params
+
 
 if sys.platform.startswith('win'):
     fix_multiproc_windows()